Rewrite the entire link creation section to be clearer and to support the complete range of attributes supported by the W3C HTML 4.01 and XHTML 1.0 standards.
One major part of the rewrite is that we now call the gr component
to generate links for glossary entries rather than reimplement a
partial version of the gr code.

Add significant numbers of new comments to make the action clearer.

Caching code commented during testing.


Update copyright date to 2007.

Updated documentation to reflect changes to the component and to be
more complete and correct.


git-svn-id: http://svn.math.hmc.edu/shared-mason-components/trunk@59 afc87059-7b49-4410-95a1-f66f4acb5064