Avoid creating an invisible link in sidebar navigation.
When generating links for pages that were not in the top-level navigation file (i.e., any sections not appearing in the navigation bar at the top of the page), `navnavnav` would create an empty link (`<a>`), and then generate the sectional navigation as a child of the `<li>` it had created.  When a file *is* in a section with a top-level navigation presence, that `<li>` contains the name of the top-level section, and so the link has a name.

Empty links are an accessibility problem; this change helps improve our compliance.