Fix the mysterious "numbered files when publishing" problem (MATHCOMP-37).
It turns out that the problem was with the `breadcrumbs` component
rather than the `publish` script.

When assembling a link, `breadcrumbs` used the path provided by
Mason's `$m->request_comp->dir_path` method, which returns the
path for the requested component as, for example, `/path/to/dir`,
with no trailing `/`.  We then do various other things, but the
key issue is that we end up with a URL that is missing its final
slash, which means that the web server sends back a 301
(permanently moved) error with the correct link (with the trailing
slash).

The `publish` script uses `wget`, and it's happy to take that new
URL and use it to pull down the page we're looking for, but it
first pulls down the URL without the slash, and saves the
resulting 301 error message to a file that has the same name as
the desired directory.  Because there is already a directory by
that name, it creates a file with `.1` appended, and, on future
published, it increments that number, creating a new file for each
time it's run.

The fix is simple: ensure that we have a trailing slash on
our directory URLs.  There are also a couple of places where we
can remove some slashes that are now redundant.

Closes MATHCOMP-37.